setup: add support for IORING_SETUP_NO_MMAP
If this flag is set when io_uring_queue_init_params() is called, then
we'll pre-allocate ring and sqe data before calling into the kernel to
setup the ring. The kernel will then map this pre-allocated data, rather
than allocate its own.

This allows use of huge pages for the backing of the rings and sqes. If
both rings and sqes fit in a single huge page, then we just need one page
for all of it. If they don't fit, then we fall back to one huge page for
the sqes, and one for the rings. This should be improved, we can really
pack things any way we want.

Signed-off-by: Jens Axboe <axboe@kernel.dk>