syscall: Avoid repeating syscall definition for each architecture
Move __sys* function definitions to syscall.h to avoid writing syscall
in inline Assembly when adding a new syscall in the future. It also
reduces the arch dependennt code.

We wrap the inline Assembly code with __do_syscall{0..6} macros. When
adding a new architecture, we also don't need to write all syscalls,
instead, we simply write __do_syscall{0..6} macros for the targeted
architecture.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>