arch/x86: Add x86 32-bit native syscall support
This adds x86 32-bit native syscall support.

Note for __do_syscall6() macro
-------------------------------
__do_syscall6() on x86 32-bit is a mess. Both Clang and GCC
cannot use %ebp in the clobber list and "r" constraint without
`-fomit-frame-pointer`

Current implementation:

For clang (the Assembly statement can't clobber %ebp):
  1) Push %ebp to preserve its value on the stack.
  2) Move the 6th argument of syscall from memory to %ebp.
  3) Do the syscall (int $0x80).
  4) Pop %ebp to restore the %ebp value.

For GCC, fortunately it has a #pragma that can force a specific
function to be compiled with -fomit-frame-pointer, so it can use
"r"(var) where var is a variable bound to %ebp.

If you have a better approach for this, please mail me.

   To: io-uring Mailing List <io-uring@vger.kernel.org>
   Cc: Ammar Faizi <ammarfaizi2@gnuweeb.org>

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>