arch: Remove `__INTERNAL__LIBURING_LIB_H` checks
We will include the syscall.h from another place as well. This check
was added by me when adding the x86 syscalls. For aarch64 we will
include this header from lib.h but we are restricted by this check.
Let's just remove it for all archs. User shouldn't touch this code
directly anyway.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>