test: Add nolibc test
This test is used to test liburing nolibc functionality. The first use
case is test get_page_size() function, especially for aarch64 which
relies on reading /proc/self/auxv. We don't seem to have a test that
tests that function, so let's do it here.

We may add more nolibc tests in this file in the future.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>