liburing: add test programs for napi busy poll
This adds two test programs to test the napi busy poll functionality. It
consists of a client program and a server program. To get a napi id, the
client and the server program need to be run on different hosts.

To test the napi busy poll timeout, the -t needs to be specified. A
reasonable value for the busy poll timeout is 100. By specifying the
busy poll timeout on the server and the client the best results are
accomplished.

Signed-off-by: Stefan Roesch <shr@devkernel.io>
Link: https://lore.kernel.org/r/20221107175357.2733763-4-shr@devkernel.io
Cc: Facebook Kernel Team <kernel-team@fb.com>
Cc: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Cc: Jens Axboe <axboe@kernel.dk>
Cc: Jakub Kicinski <kuba@kernel.org>
Cc: Olivier Langlois <olivier@trillion01.com>
Cc: io-uring Mailing List <io-uring@vger.kernel.org>
Cc: netdev Mailing List <netdev@vger.kernel.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>