nolibc: Stop defining `malloc()` and `free()` as a global function
With libc, calling malloc() and free() are just a function call to a
libc function. Without libc, malloc() and free() are just a syscall
(mmap and munmap).

There is no real benefit to defining these functions in nolibc.c.
Invoking them directly inline is small, and it reduces the code size.

Also, stop calling them uring_malloc() and uring_free(). Just define
malloc() and free() as a macro. This way, we have the same name as
libc, and no conflict is introduced.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>